requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.0.0
pytest-mock==3.12.0
Pillow==10.2.0 